Move api routes to their own folder
