+++ functions/src/api/routes/sponsors/sponsors.http
